+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 - [Feature: Support custom rate limit reject response info](https://github.com/Tencent/spring-cloud-tencent/pull/153)
 - [Feature: Remove spring-javaformat-maven-plugin](https://github.com/Tencent/spring-cloud-tencent/pull/151)
 - [feat:optimize config server address.](https://github.com/Tencent/spring-cloud-tencent/pull/149)
+- [feat: override recover router config](https://github.com/Tencent/spring-cloud-tencent/pull/159)
 - [feat:refactor loadbalancer module as a basic module for router and circuit breaker.](https://github.com/Tencent/spring-cloud-tencent/pull/155)